wireframe functionality update
Made it so you don't have to be logged in to see the main page. App now identifies if you are logged in for both the main page and login page and acts accordingly.

Not fully done for the search cards page, that is, you can't visit the cards page without being logged in.