Big Update
both login and register modals set up and working with the api, we can log in and register! for serious!

login page now deleted, login javascript functionality carried over to loginModal

the modals dynamically appear on both the root and games page, depending on if you are logged in or out

when you log out you now stay on the page you logged out on

todo: implement password retype input and string compare check the two before registering

enforce email formatting before registering

email verification

when you sign in you go only to games page right now (didn't fix this because it's the only way to get to the games page other than manually typing in the url.