checkbox search fully implemented
everything should be working (platform and genre), see if it can be broken!  We can mix and match as much as we want before searching. Not all genres or platforms available.

made a change in the backend to make genre match platform in searchAllGames because genre was only searching the array in order of matching values (ex: if we put in ["RPG", "FPS"] we'd only get a result if it was specifically in that order.